requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
pydantic==2.7.0
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
python-jose==3.3.0
python-multipart==0.0.9
pyjwt==2.8.0